Move `_Batch` and `_BatchManager`
